# 8264777: Overload optimized FileInputStream::readAllBytes