--- a/src/java.base/share/classes/java/io/FileInputStream.java
+++ b/src/java.base/share/classes/java/io/FileInputStream.java
@@ -275,13 +275,20 @@ public class FileInputStream extends InputStream
     }
 
     public byte[] readAllBytes() throws IOException {
-        long size = length() - position();
-        if (size > (long) Integer.MAX_VALUE)
-            throw new OutOfMemoryError("Required array size too large");
-        if (size <= 0L)
-            return new byte[0];
+        long length = length();
+        int capacity;
+        if (length != 0L) {
+            long position = position();
+            long size = length - position;
+            if (size > (long)Integer.MAX_VALUE)
+                throw new OutOfMemoryError("Required array size too large");
+            if (size <= 0L)
+                return new byte[0];
+            capacity = (int)(length - position);
+        } else {
+            capacity = DEFAULT_BUFFER_SIZE;
+        }
 
-        int capacity = (int)size;
         byte[] buf = new byte[capacity];
 
         int nread = 0;
@@ -311,11 +318,18 @@ public class FileInputStream extends InputStream
     public byte[] readNBytes(int len) throws IOException {
         if (len < 0)
             throw new IllegalArgumentException("len < 0");
-        long size = length() - position();
-        if (size <= 0L)
+        if (len == 0L)
             return new byte[0];
 
-        int capacity = (int)Math.min(len, size);
+        long length = length();
+        int capacity;
+        if (length == 0L) {
+            capacity = Math.min(len, DEFAULT_BUFFER_SIZE);
+        } else {
+            capacity = (int)Math.min(len, length - position());
+            if (capacity <= 0L)
+                return new byte[0];
+        }
         byte[] buf = new byte[capacity];
 
         int remaining = capacity;