--- a/src/java.base/share/native/libjava/FileInputStream.c
+++ b/src/java.base/share/native/libjava/FileInputStream.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2021, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -70,6 +70,39 @@ JNIEXPORT jint JNICALL
 Java_java_io_FileInputStream_readBytes(JNIEnv *env, jobject this,
         jbyteArray bytes, jint off, jint len) {
     return readBytes(env, this, bytes, off, len, fis_fd);
+}
+
+JNIEXPORT jlong JNICALL
+Java_java_io_FileInputStream_length(JNIEnv *env, jobject this) {
+
+    FD fd;
+    jlong length = jlong_zero;
+
+    fd = getFD(env, this, fis_fd);
+    if (fd == -1) {
+        JNU_ThrowIOException(env, "Stream Closed");
+        return -1;
+    }
+    if ((length = IO_GetLength(fd)) == -1) {
+        JNU_ThrowIOExceptionWithLastError(env, "GetLength failed");
+    }
+    return length;
+}
+
+JNIEXPORT jlong JNICALL
+Java_java_io_FileInputStream_position(JNIEnv *env, jobject this) {
+    FD fd;
+    jlong ret;
+
+    fd = getFD(env, this, fis_fd);
+    if (fd == -1) {
+        JNU_ThrowIOException(env, "Stream Closed");
+        return -1;
+    }
+    if ((ret = IO_Lseek(fd, 0L, SEEK_CUR)) == -1) {
+        JNU_ThrowIOExceptionWithLastError(env, "Seek failed");
+    }
+    return ret;
 }
 
 JNIEXPORT jlong JNICALL