Please consider this request to override the `java.io.InputStream` methods `readAllBytes()` and `readNBytes(int)` in `FileInputStream` with more performant implementations. The method overrides attempt to read all requested bytes into a single array of the required size rather than composing the result from a sequence of smaller arrays. An example of the performance improvements is as follows.

**readAllBytes**
Before
```
Benchmark                                    (length)   Mode  Cnt      Score     Error  Units
ReadAllBytes.readAllBytesFileInputStream      1000000  thrpt   20   2412.031 ±   7.309  ops/s
ReadAllBytes.readAllBytesFileInputStream     10000000  thrpt   20    212.181 ±   0.369  ops/s
ReadAllBytes.readAllBytesFileInputStream    100000000  thrpt   20     19.860 ±   0.048  ops/s
ReadAllBytes.readAllBytesFileInputStream   1000000000  thrpt   20      1.298 ±   0.183  ops/s
```
After
```
Benchmark                                    (length)   Mode  Cnt      Score     Error  Units
ReadAllBytes.readAllBytesFileInputStream      1000000  thrpt   20   8218.473 ±  43.425  ops/s
ReadAllBytes.readAllBytesFileInputStream     10000000  thrpt   20    302.781 ±   1.273  ops/s
ReadAllBytes.readAllBytesFileInputStream    100000000  thrpt   20     22.461 ±   0.084  ops/s
ReadAllBytes.readAllBytesFileInputStream   1000000000  thrpt   20      1.502 ±   0.003  ops/s
```

**readNBytes**

`N = file_size / 2`

Before
```
Benchmark                                    (length)   Mode  Cnt      Score     Error  Units
ReadAllBytes.readHalfBytesFileInputStream     1000000  thrpt   20   5585.500 ± 153.353  ops/s
ReadAllBytes.readHalfBytesFileInputStream    10000000  thrpt   20    436.164 ±   1.104  ops/s
ReadAllBytes.readHalfBytesFileInputStream   100000000  thrpt   20     40.167 ±   0.141  ops/s
ReadAllBytes.readHalfBytesFileInputStream  1000000000  thrpt   20      3.291 ±   0.211  ops/s
```

After
```
Benchmark                                    (length)   Mode  Cnt      Score     Error  Units
ReadAllBytes.readHalfBytesFileInputStream     1000000  thrpt   20  15463.210 ±  66.045  ops/s
ReadAllBytes.readHalfBytesFileInputStream    10000000  thrpt   20    561.752 ±   0.951  ops/s
ReadAllBytes.readHalfBytesFileInputStream   100000000  thrpt   20     45.043 ±   0.102  ops/s
ReadAllBytes.readHalfBytesFileInputStream  1000000000  thrpt   20      4.629 ±   0.035  ops/s
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8264777](https://bugs.openjdk.java.net/browse/JDK-8264777): Overload optimized FileInputStream::readAllBytes


### Reviewers
 * [Daniel Fuchs](https://openjdk.java.net/census#dfuchs) (@dfuch - **Reviewer**) ⚠️ Review applies to 4fae0209f7523a8b2f6f04abb451be913c624881
 * [Alan Bateman](https://openjdk.java.net/census#alanb) (@AlanBateman - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/3845/head:pull/3845` \
`$ git checkout pull/3845`

Update a local copy of the PR: \
`$ git checkout pull/3845` \
`$ git pull https://git.openjdk.java.net/jdk pull/3845/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 3845`

View PR using the GUI difftool: \
`$ git pr show -t 3845`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/3845.diff">https://git.openjdk.java.net/jdk/pull/3845.diff</a>

</details>