@AlanBateman 2021-05-16
> Any more comments on this? Thanks.

I think you've addressed the obvious bugs now but I'm still nervous that the lseek will fail for some special devices (time will tell).

It would be good to do some cleanup on the test before you integrate this. If you change it to use try-with-resources will it will avoid leaving the file open when the test fails. Also would be good to add a test for readNBytes(0) for the empty-file case. There's a typo in one of the exception messages ("expecte").